# swagger.yaml
openapi: 3.0.3
info:
  title: SENG2021 Microservice API
  version: 1.0.0
  description: API documentation for DespatchAdvice, ReceiptAdvice, FulfilmentCancellation, and OrderAdjustment.
servers:
  - url: http://localhost:3000
    description: Local server
security:
  - BearerAuth: []
paths:
...
  /api/despatch-advices:
    post:
      summary: Create a new Despatch Advice (UBL 2.1 XML)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DespatchAdviceRequest'
      responses:
        '201':
          description: Created successfully
          content:
            application/xml:
              schema:
                type: string
                example: "<DespatchAdvice>...</DespatchAdvice>"
        '400':
          description: Validation error
        '401':
          description: Unauthorized - missing or invalid token
        '403':
          description: Forbidden - wrong role
        '422':
          description: Invalid quantity
  /api/despatch-advices/{dispatchAdviceId}:
    get:
      summary: Get a Despatch Advice by ID (UBL XML)
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: dispatchAdviceId
          required: true
          schema:
            type: string
            example: DA-12345
      responses:
        '200':
          description: Found
          content:
            application/xml:
              schema:
                type: string
                example: "<DespatchAdvice>...</DespatchAdvice>"
        '401':
          description: Unauthorized - missing or invalid token
        '404':
          description: Despatch Advice not found

  # -------------------- ReceiptAdvice --------------------
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Pass your partyId as the Bearer token (e.g. "Bearer SUP-01")
  schemas:
...
    DespatchAdviceRequest:
      type: object
      required:
        - despatchParty
        - deliveryParty
        - dispatchDate
        - items
      properties:
        externalRef:
          type: string
          example: REF-001
        despatchParty:
          type: object
          required:
            - partyId
            - name
          properties:
            partyId:
              type: string
              example: SUP-01
            name:
              type: string
              example: Supplier Alpha
        deliveryParty:
          type: object
          required:
            - partyId
            - name
          properties:
            partyId:
              type: string
              example: WH-02
            name:
              type: string
              example: Warehouse Beta
        dispatchDate:
          type: string
          format: date-time
          example: '2026-03-15T06:00:00Z'
        expectedDeliveryDate:
          type: string
          format: date-time
          example: '2026-03-16T06:00:00Z'
        items:
          type: array
          items:
            type: object
            required:
              - sku
              - description
              - quantity
              - uom
            properties:
              sku:
                type: string
                example: ITEM-A
              description:
                type: string
                example: Blue Widget
              quantity:
                type: integer
                example: 5
              uom:
                type: string
                example: EA

    # -------------------- ReceiptAdvice --------------------
...